refactor(helpme): narrow client boundary to ClickableTicketRow
- Move row-click handler into a small client wrapper so ticket-list
  stays server-rendered (cells, chips, date formatter no longer ship to
  the browser).
- Honor modifier keys (Cmd/Ctrl/Shift/Alt) on row click so they fall
  through to native anchor behavior (open in new tab/window).
- Skip row handler when the click lands on an anchor — lets <Link> own
  its own navigation without needing stopPropagation.
- Drop the redundant onMouseEnter router.prefetch; next/link already
  prefetches the subject link when it enters the viewport.
